move karma.js and tests to top level
